backend/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt
django-filter>=23.5
drf-spectacular>=0.26.0
drf-spectacular-sidecar>=2025.5.0
django-environ>=0.11.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.0.0
textblob>=0.19.0
langchain>=0.0.352
langchain-community>=0.0.10
openai>=1.6.1
python-dotenv>=1.0.0
tiktoken>=0.5.2
ics>=0.7.2
pytz>=2024.1
weasyprint>=65.1
scikit-learn>=1.3.0
